(drivers) imx6_net: Refactor
Move declarations and macros to separate file
